MFC r342952:
Add Linux compatibility support for `SC_NPROCESSORS_{CONF,ONLN}` as `_SC_NPROCESSORS_{CONF,ONLN}`

The goal of this change is to make it easier to use getconf to query
the number of available processors.

Sadly it's unclear per POSIX, which form (with a preceding _ or
lacking it) is correct. I will bring this up on the Austin Group list so
this point is clarified for implementors that might rely on this getconf
variable in future POSIX spec versions.

This is something I noticed when trying to import GoogleTest to FreeBSD
as one of the CI scripts uses this variable on Linux.

Approved by:    emaste (mentor; implicit: https://reviews.freebsd.org/D19527)
Differential Revision: https://reviews.freebsd.org/D19528


git-svn-id: https://svn.freebsd.org/base/stable/12@345016 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f