MFC 331417,331597: Add the cryptocheck test tool.
331417:
Bring in JHB's cryptocheck tool

It can be used to validate basic algorithm correctness on a variety of inputs,
by comarison to openssl.

While here, add some sanity to the crypto/Makefile.

The tool may not be perfect, but getting it in tree where collaboration can
happen is a nice first step.  The pace of development outside of svn seems
to have slowed down mid-2017.

331597:
Update the license to note my work on cryptocheck was sponsored.

Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/11@345028 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f