MFC r343998:
Fix off-by-one error in BERI virtio driver

The hardcoded ident is exactly 20 bytes long but sprintf adds terminating zero,
so there is one byte written out of array bounds.As a fix use strncpy it
appends \0 only if space allows and its behavior matches virtio spec:

When VIRTIO_BLK_T_GET_ID is issued, the device identifier, up to 20 bytes, is
written to the buffer. The identifier should be interpreted as an ascii string.
It is terminated with \0, unless it is exactly 20 bytes long.

PR:		202298
Reviewed by:	br
Differential Revision:	https://reviews.freebsd.org/D18852


git-svn-id: https://svn.freebsd.org/base/stable/12@345462 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f