loader: mod_loadkld() error: we previously assumed 'last_file' could be null
The last_file variable is used to reset the loadaddr variable back to original
value; however, it is possible the last_file is NULL, so we can not blindly
trust it. But then again, we can just save the original loadaddr and use
the saved value for recovery.

MFC after:	1w


git-svn-id: https://svn.freebsd.org/base/head@346002 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f