Update and clarify pflog man page
The pflog0 interface is created when the module is loaded, this can
be triggered by pf and pflogd being enabled or by kldloading the module.

By default the interface would be pflog0, add the ifconfig stage of the
example to make this example clearer.

Reviewed by:	kp, bz, bcr, jtl, 0mp
Approved by:	jtl (mentor), bz (mentor)
Differential Revision:	https://reviews.freebsd.org/D19861


git-svn-id: https://svn.freebsd.org/base/head@346274 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f