MFC: r345992, r346087 Add INET6 support for the upcalls to the nfsuserd daemon.
The kernel code uses UDP to do upcalls to the nfsuserd(8) daemon to get
updates to the username<->uid and groupname<->gid mappings.
A change to AF_LOCAL last year had to be reverted, since it could result
in vnode locking issues on the AF_LOCAL socket.
This patch adds INET6 support and the required #ifdef INET and INET6
to the code.


git-svn-id: https://svn.freebsd.org/base/stable/11@346463 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f